Added Inventory. Added ability to pickup, drop and replace items in inventory. Reworked item cycling for dynamic spell usage.
Fixed an issue where stone cannon would collide with the hitbox of dispel magic area and impact it.

Needs rework: Inventory spell casting and item usage needs to be more dynamic as its implementation is poor. Needs Database for Items so that its easier to filter and use them. Low priority because its funtional with current setup.